# api.yml
openapi: 3.0.3
info:
  title: CapitalCoa API
  version: 1.0.0
  description: >
    REST API for CapitalCoa PWA: manage assets, signals, trades, backtests,
    push subscriptions and audit logs. Uses JWT bearer auth.
servers:
  - url: https://api.yourdomain.com
    description: Production server
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /api/auth/login:
    post:
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: bearer
  /api/auth/logout:
    post:
      summary: Admin logout (invalidate token)
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logged out
...